Trader: bugの温床になるため、float32の使用を取りやめ

+++ models/oanda_py_client.py
@@ -432,11 +432,11 @@ class OandaPyClient():
 
         candle = pd.DataFrame.from_dict([row['mid'] for row in response['candles']])
         candle = candle.astype({
-            # INFO: 'float64' より早くなるらしいが...
-            'c': 'float32',
-            'l': 'float32',
-            'h': 'float32',
-            'o': 'float32'
+            # INFO: 'float32' の方が速度は早くなるが、不要な小数点4桁目以下が出現するので64を使用
+            'c': 'float64',
+            'l': 'float64',
+            'h': 'float64',
+            'o': 'float64'
         })
         candle.columns = ['close', 'high', 'low', 'open']
         candle['time'] = [row['time'] for row in response['candles']]
